refactor: change variable type to long
Refactored all models to have their IDs use Long instead of int

+++ src/main/java/es/daw2/fct_fct/repositorio/RepositorioAlumno.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import es.daw2.fct_fct.modelo.Alumnos;
 
-public interface RepositorioAlumno extends CrudRepository<Alumnos,Integer>{
+public interface RepositorioAlumno extends CrudRepository<Alumnos, Long>{
 
 }
